Prep for PCM Experiments
Changes are made for a simple scoping experiment of PCM in a office
space environment. There will be one sensor hooked up to the DAQ that
will take flow rate and temperature readings. Temperature will be hooked
up to analog 0 and the flow rate will be hooked up to analog 1.