Datasheet
Added the datasheet for the Grundfos VFS 1-20 flow sensor that will be
used for the experiment. There was a test run done, however the files
were too large to be stored on GitHub and was transferred to the Toshiba
work computer in the office. The DAQ settings have been modified to take
less samples at a lower frequency to allow for longer runs and smaller
file sizes.